chore: 더미 이미지 데이터 분리
- 더미 이미지 생성 함수 `generateRandomImageUrls`를 `DummyData` 객체로 이동
